JustChat/requirements.txt
# Web框架
flask==2.3.3
flask-cors==4.0.0

# 微信接入
itchat==1.3.10

# AI对话
openai==1.3.0

# 配置管理
python-dotenv==1.0.0

# 加密
cryptography==41.0.0

# 数据库
# sqlite3是Python内置模块,无需安装

# PDF处理
PyPDF2==3.0.1

# 中文分词
jieba==0.42.1

# 测试
pytest==7.4.2
pytest-cov==4.1.0

# 部署
gunicorn==21.2.0
